Moved backend source from backend/backend -> backend/src
As requested in issue #4
